requirements.txt
PyCrypto>=2.6.1
python-gnutls>=2.0.1
